catch footgun: duplicate groups in input led to incomplete output.
Repeatedly specifying the same group in the input would cause that groups'
tempfile to be used multiple times, thus re-setting the recursive joining
process to the state of when the group was first used.
this reset causes intermediate steps to be lost.

Prevent this by deduplicating the group list before starting.